requirements/base.txt
Django==2.0.7
django-filter==2.0.0
django-service-objects==0.4.0
djangorestframework==3.8.2
psycopg2==2.7.5
flake8==3.5.0
python-dotenv==0.8.2
django-cors-headers==2.4.0
PyJWT==1.6.4
requests==2.19.1
